JDK11-52: Warning before deleting all finished translation concepts.

+++ rest/src/main/webapp/app/component/translationTable/translationDirective.js
@@ -457,10 +457,12 @@ tsApp
 
               // Removes all translation concepts
               $scope.removeAllTranslationConcepts = function(translation) {
-                translationService.removeAllTranslationConcepts(translation.id).then(
-                  function(data) {
-                    translationService.fireTranslationChanged(translation);
-                  });
+                if ($window.confirm("Are you sure you want to remove the translation concepts in " + translation.name + "?")) {
+                  translationService.removeAllTranslationConcepts(translation.id).then(
+                    function(data) {
+                      translationService.fireTranslationChanged(translation);
+                    });
+                }
               };
 
               // Convert date to a string
